% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-table-one-tests.R
\name{calc_table_one_tests}
\alias{calc_table_one_tests}
\title{Return variables from table one to look into}
\usage{
calc_table_one_tests(table_one)
}
\arguments{
\item{table_one}{A tibble result from \code{create_tidy_table_one}}
}
\value{
A list
}
\description{
Given the results of the \code{create_tidy_table_one} this will give you a list
with the variable names for those that were, for example, non-normal by
Shapiro-Wilkes test or that had unequal variance by Barlett's test.
}
\examples{
library(dplyr)

tab1 <- create_tidy_table_one(data = pbc_mayo,
                              strata = "trt",
                              .vars = c("time",
                                        "status",
                                        "trt",
                                        "age",
                                        "sex",
                                        "ascites",
                                        "hepato",
                                        "spiders",
                                        "edema",
                                        "bili",
                                        "chol",
                                        "albumin",
                                        "copper",
                                        "alk_phos",
                                        "ast",
                                        "trig",
                                        "platelet",
                                        "protime",
                                        "stage"))

dplyr::glimpse(tab1)

calc_table_one_tests(tab1)
}
