% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-tidy-table-one_no-strata.R
\name{create_tidy_table_one_no_strata}
\alias{create_tidy_table_one_no_strata}
\title{Tidy table one with no strata}
\usage{
create_tidy_table_one_no_strata(
  data,
  strata = NULL,
  vars,
  na_level = "(Missing)",
  b_replicates = 2000,
  ...
)
}
\arguments{
\item{data}{A data frame or tibble containing the variables to be summarized.}

\item{strata}{NULL}

\item{vars}{Character vector of the variable names to be summarized. If
empty, then all variables in the given data frame are used.}

\item{na_level}{Character string of the text to replace \code{NA} in the strata
variable, if any exist.}

\item{b_replicates}{an integer specifying the number of replicates used in
the Monte Carlo test for Fisher's Exact test and Chi-square test.}

\item{...}{Additional arguments. Not used.}
}
\value{
A tibble with the following results
\describe{
\item{strata}{Level of the stratifying variable}
\item{var}{Variable/column name}
\item{n}{Number of records}
\item{n_distinct}{Numer of distinct values}
\item{complete}{Number of non-missing observations}
\item{missing}{Number of missing observations}
\item{mean}{Mean}
\item{sd}{Standard deviation}
\item{p0}{Minimum}
\item{p25}{25th percentile}
\item{p50}{Median}
\item{p75}{75th percentile}
\item{p100}{Maximum}
\item{cv}{Coefficient of variation}
\item{shapiro_test}{Shapiro-Wilkes test: p-value}
\item{ks_test}{Kolmogorov-Smirnov test: p-value}
\item{ad_test}{Anderson-Darling test for normality: p-value}
\item{level}{Level of the variable}
\item{n_level}{Total number in the variable's group}
\item{n_strata}{Total number in the variable group and strata}
}
}
\description{
Creates a tidy data frame of the results that can go into a "Table 1" of
summary descriptive statistics of a study sample. Inspiration for this is owed
to the \code{tableone} package by Kazuki Yoshida.
}
\examples{
library(dplyr)

tab1 <- create_tidy_table_one_no_strata(data = pbc_mayo,
                                        strata = NULL,
                                        vars = c("time",
                                                 "status",
                                                 "trt",
                                                 "age",
                                                 "sex",
                                                 "ascites",
                                                 "hepato",
                                                 "spiders",
                                                 "edema",
                                                 "bili",
                                                 "chol",
                                                 "albumin",
                                                 "copper",
                                                 "alk_phos",
                                                 "ast",
                                                 "trig",
                                                 "platelet",
                                                 "protime",
                                                 "stage"))

dplyr::glimpse(tab1)


library(ggplot2)  # diamonds data set

tab2 <- create_tidy_table_one_no_strata(data = diamonds,
                                        vars = c("carat",
                                                 "cut",
                                                 "color",
                                                 "clarity",
                                                 "depth",
                                                 "table",
                                                 "price"))

dplyr::glimpse(tab2)
}
