% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_one_helpers.R
\name{get_var_info}
\alias{get_var_info}
\title{Get info about variables in a data frame}
\usage{
get_var_info(data, .vars = NULL)
}
\arguments{
\item{data}{A data.frame or tbl_df}

\item{.vars}{A character string of variable/column names. If empty, all
columns are used}
}
\value{
A tbl_df with colums for
\describe{
\item{var}{name of the variable/column}
\item{level}{factor level if applicable, character levels too}
\item{class}{variable class}
\item{var_type}{variable type: categorical or continuous}
}
}
\description{
Mostly a helper function for other tidy_table_one functions. The function
takes a data frame or tibble and returns a descriptive tibble with the
variable names, class, binary type (categorical or continuous), and the
factor levels.
}
\examples{
library(dplyr)
library(ggplot2)
lapply(diamonds, class)
get_var_info(data = diamonds,
             .vars = c("carat",
                       "cut",
                       "color",
                       "clarity",
                       "depth",
                       "table",
                       "price")) \%>\%
  print(n = Inf)


get_var_info(data = pbc_mayo)

get_var_info(data = pbc_mayo,
             .vars = c("status", "sex", "stage", "age"))
}
