% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-tidytableone-no-strata.R,
%   R/create-tidytableone.R, R/deprecations.R
\name{create_tidy_table_one_no_strata}
\alias{create_tidy_table_one_no_strata}
\alias{create_tidy_table_one_no_strata_checkbox}
\alias{create_tidy_table_one}
\alias{create_tidy_table_one_checkbox}
\alias{tidytableone-deprecated}
\title{Deprecated functions in tidytableone}
\usage{
create_tidy_table_one_no_strata(...)

create_tidy_table_one_no_strata_checkbox(...)

create_tidy_table_one(...)

create_tidy_table_one_checkbox(...)
}
\description{
The following functions are deprecated and will be removed in a future release.
Use the replacements instead:
}
\details{
\itemize{
\item \code{\link[=create_tidytableone]{create_tidytableone()}} replaces \code{\link[=create_tidy_table_one]{create_tidy_table_one()}}.
\item \code{\link[=create_tidytableone_checkbox]{create_tidytableone_checkbox()}} replaces \code{\link[=create_tidy_table_one_checkbox]{create_tidy_table_one_checkbox()}}.
\item \code{\link[=create_tidytableone_no_strata]{create_tidytableone_no_strata()}} replaces \code{\link[=create_tidy_table_one_no_strata]{create_tidy_table_one_no_strata()}}.
\item \code{\link[=create_tidytableone_no_strata_checkbox]{create_tidytableone_no_strata_checkbox()}} replaces \code{\link[=create_tidy_table_one_no_strata_checkbox]{create_tidy_table_one_no_strata_checkbox()}}.
}
}
\keyword{internal}
