% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytableone-package.R
\docType{package}
\name{tidytableone-package}
\alias{tidytableone}
\alias{tidytableone-package}
\title{tidytableone: Tidy construction of "Table 1"}
\description{
A tidy approach to creating the "Table 1" of study characteristics common
in biomedical research. The two main functions are
\code{\link[=create_tidytableone]{create_tidytableone()}} (which produces a tibble of summary statistics and
hypothesis-test results) and \code{\link[=adorn_tidytableone]{adorn_tidytableone()}} (which formats that
tibble into a printable Table 1).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emilelatour/tidytableone}
  \item Report bugs at \url{https://github.com/emilelatour/tidytableone/issues}
}

}
\author{
\strong{Maintainer}: Emile Latour \email{latour@ohsu.edu}

Authors:
\itemize{
  \item Emile Latour \email{latour@ohsu.edu}
}

}
\keyword{internal}
