% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytableone.R
\docType{package}
\name{tidytableone}
\alias{tidytableone}
\title{tidytableone: Tidy construction of "Table 1"}
\description{
A tidy approach to creating the "Table 1" of study characteristics common in
biomedical research.
}
\examples{
# Example usage:
library(tidytableone)

}
