% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytableone.R
\docType{package}
\name{tidytableone-package}
\alias{tidytableone}
\alias{tidytableone-package}
\title{tidytableone: Tidy construction of "Table 1"}
\description{
A tidy approach to creating the "Table 1" of study characteristics common in
biomedical research.
}
\examples{
# Example usage:
library(tidytableone)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emilelatour/tidytableone}
  \item Report bugs at \url{https://github.com/emilelatour/tidytableone/issues}
}

}
\author{
\strong{Maintainer}: Emile Latour \email{latour@ohsu.edu}

}
\keyword{internal}
